refactor: refactor validateMediaAndText folder
