+++ src/utils/validateMediaAndText/mocks/mockFiles.ts
@@ -1,0 +1,17 @@
+const _10_MB_FILE = 1;
+
+export const mockFileImage = new File(
+  [new ArrayBuffer(_10_MB_FILE)],
+  'mockfile.jpg',
+  {
+    type: 'image/jpeg',
+  }
+);
+
+export const mockFileVideo = new File(
+  [new ArrayBuffer(_10_MB_FILE)],
+  'mockfile.mp4',
+  {
+    type: 'video/mp4',
+  }
+);
