+++ src/utils/validateMediaAndText/validadeText/validateText.ts
@@ -1,0 +1,3 @@
+export function validateTextLength(text: string, limitLength: number): boolean {
+  return text.length <= limitLength;
+}
