+++ src/utils/validateMediaAndText/validateMedia/validateMedia.ts
@@ -1,0 +1,3 @@
+export function validateMediaSize(file: File, limitSize: number): boolean {
+  return file.size <= limitSize;
+}
